refactor: 에이스 합계 계산 책임을 Rank에서 Score로 변경
